+++ src/twitch/twitch.ts
@@ -1,3 +1,4 @@
+import { initExpressApp } from '../express.js'
 import { createAuthAndApiClient } from './twitchApi.js'
 import { initTwitchChat } from './twitchChat.js'
 import { initTwitchEventSub } from './twitchEventSub.js'
@@ -5,7 +6,8 @@ import { initTwitchOAuthServer } from './twitchOAuth.js'
 
 export async function initTwitch() {
 	const { authProvider, apiClient, helixUsers } = await createAuthAndApiClient()
-	const expressApp = await initTwitchOAuthServer()
+	const expressApp = await initExpressApp()
+	initTwitchOAuthServer(expressApp)
 	await Promise.all([
 		initTwitchChat(authProvider, helixUsers.bot),
 		initTwitchEventSub({
